requirements.txt
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.5.1
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
